Rename "theme.extension" to "theme.parent"
This makes it so any theme can be a clone of another theme with minor modifications.

Expected values are no longer "styled" or "unstyled", but "_styled" or "_unstyled" if it references themes within Liferay core. An empty value signifies that a theme is not a clone of any other theme. A non empty file other than "_styled" or "_unstyled" signifies that it wants a directory of that theme to clone from.

git-svn-id: svn://svn.liferay.com/repos/public/plugins/trunk@14093 05bdf26c-840f-0410-9ced-eb539d925f36