LPS-45271 It is no longer necessary to have different owner types to represent defaults.
Previously this was necessary to avoid collisions, but since we are now using serviceName as the key for the settings of services and portletId as the key for the settings of portlet instances there is no risk of collisions. Considering this, we think it's better to simplify and reduce the need for extra owner types.