LPS-47337 Resolve dependencies from the SDK and from the module
Since we are always executing resolve before resources there is no need
to pass file or transitive attributes. These are not taken into account
and can be misleading since it may appear that we have path executions
in which ivy:resolve has not been called.

We don't need to specify conf either since we are only resolving test in
the previous calls to resolve.