LPS-51413 Fix race condition - not related
It's unlikely that "shutdown" would ever be called twice, however it's
possible for "request" to be called from several simultaneous threads at
the same time shutdown is invoked. If "_stopped" is just an ordinary
field, threads serving requests might not see its current value and
therefore not be aware that the Solr server has already been shut down.