LPS-52790 cannot use PrevAndNext method
Because there can be multiple articles with the same priority, we cannot
reliably get the next and previous article using a single SQL query.

The only way to guarantee that the returned next and previous articles
are consistent with the order presented in the navigation is to scan the
list of returned articles.