LPS-53289 always use request parameters
The kbArticle saved in the attribute is selected using different rules
depending on the portlet (e.g. the Display portlet uses a quite complex
article selection algorithm).

When moving an article, the IDs for that element are already
provided as request parameters (and are always correct, as opposed to
the kbArticle in the attribute).