LPS-53649 Invalidate cache on move
When a file entry is moved, its path changes. If we don't remove the
element from the cache, any further requests for this object would
return one with the old path. This caused the DM to keep references to
the old path and failing to find the file entry, as Sharepoint uses this
path as a primary key.