LPS-79468 Fix it in another way: eliminate the 'searchCount' call (which results translating the query twice) by adapting the core changes of LPS-72199 to SolrIndexSearcher from 70x:
liferay/liferay-portal@858d1aff1bf2d0f7c05ab518c5b5526c598905e6 :

"Improve performance and stability by using 'guess and check' vs 'check and guess' pattern. It's a guess currently because the state could change between the count and the search. It's faster because we don't have to perform the count before we do the search and the normal case is guessing correctly."

https://issues.liferay.com/browse/LPS-79468